Add Custom Middleware to the UI to Configure HttpClient's Header
The access token is set or removed in httpClient's default request header in the SettingHttpClientHeaderMiddleware. There is no longer need to manually input token to the httpClient header each time the request is sent.
Adds ApiHelper class to the UI library, which creates one apiClient, which is used to call HAAPI. No longer need to create a HttpClient each time before calling the api.